stt/requirements.txt
TTS
torchaudio
fastapi
uvicorn
faster_whisper
numpy
websockets
scipy
aiohttp
sounddevice
webrtcvad
livekit
livekit-api
